add new command folder for audio based commands and move purr.js to said folder
