--- .env.sample
@@ -1,7 +1,0 @@
-TOKEN=botTokenHere
-CHANNEL_ID=yourChannelID
-DB_USER=yourUserName
-DB_PASSWORD=yourUserPassword
-DB_NAME=yourDBName
-COLLECTION_NAME=yourCollectionName
-CLUSTER_NAME=yourClusterName
