+++ util/detectLink.js
@@ -3,5 +3,10 @@
  * @param {String} string string
  */
 module.exports = function(string) {
-
+    if (string.includes("http://") || string.includes("https://")) {
+        return true;
+    } 
+    else {
+        return false;
+    }
 }
