+++ util/react.js
@@ -1,8 +1,8 @@
 //checks if message contains one of two strings and is in the desired channel then reacts to that message with an emoji
-module.exports = function(message, channelID) {
+module.exports = function(message, channelID, emote) {
     if ((message.content.includes('odie') || message.content.includes('jobie')) && message.channel.id == channelID) {
         try {
-            message.react('505132652437700642');            
+            message.react(emote);            
         }
         catch(e) {
             console.log(e);
